requirements.txt
aiohttp==3.8.1
asyncpg==0.27.0
Pillow==9.3.0
beautifulsoup4==4.11.1
psycopg2-binary==2.9.5